requirements.txt
setuptools==74.1.2
python-frontmatter==1.1.0